First GUI protoype
Added the first prototype for the GUI, with basic functionality,
though no reason to use tools until the project can be set up,
the images loaded, and the manual picker is ready.

Lots of bugs and unfinished features. Need to modify the save dialog
to allow for entering filenames with keyboard... Still substantially
more work to perform.

	added:  debug/
	        debug/default_picker_save.tpp
	        ui/
	        ui/test_1.ui
	        term-peeper.py

	modified:  roadmap.md