Added Data Download Dialog
Added the ability to download MODIS scenes through a menu in
the main window. This utilizes pymodis to retrieve and
download the hdf and xml files. Currently the product cannot be
chosen, and it is hardcoded as MOD09GA. Conversion, projection, etc
are not yet added. Progress bar for downloading images is not
visible during downloading. Might need to subclass QDialog to get
the progress bar working as planned.

modified:   term-peeper.py
new file:   ui/data_download_dialog.ui
modified:   ui/test_1.ui