Aqua and Terra Download and File Interleaving
Fulfills request #3

Aqua and Terra can now be downloaded in the Modis Download Dialog
In the project file creation dialog, data from these two satellite platforms
are interleaved upon choosing HDF files. There is now also an auto sort
button at the bottom of the file list to help with data from multiple
imports.

The sort order is such that for two adjacent images in the file queue, assuming
the modis HDF retain their original filename conventions, a terra scene (MOD)
will be infront of an aqua scene (MYD) if the two fall on the same Julian date.

Replaced the copyright notification on the UI files

	modified:   term-peeper/modis.py
	modified:   term-peeper/termpeeperIO.py
	modified:   ui/data_download_dialog.ui
	modified:   ui/main_window.ui
	modified:   ui/new_project_wizard_page_1.ui